requirements.txt
Django==5.2.9
djangorestframework==3.16.1
Pillow==10.4.0
gunicorn==21.2.0
whitenoise==6.7.0

